Add machine and runtime information
Split atsamd21 to atsamd21-e18/g18
